Sim - refactor ui by breaking out forms into separate files
